Fix incomplete project properties
I found that in a certain case, I wasn't getting all of the preprocessor
definitions that were present:

In a property sheet:

  <ItemDefinitionGroup>
    <ClCompile>
       <PreprocessorDefinitions>NOMINMAX;%(PreprocessorDefinitions)</PreprocessorDefinitions>
    </ClCompile>
  <ItemDefinitionGroup>
  <ItemDefinitionGroup Condition="'$(Configuration)'=='Debug'">
    <ClCompile>
      <PreprocessorDefinitions>_DEBUG;%(PreprocessorDefinitions)</PreprocessorDefinitions>
    </ClCompile>
  </ItemDefinitionGroup>

In a project including the property sheet:

  <ItemDefinitionGroup Condition="'$(Configuration)'=='Debug'">
    <ClCompile>
      <PreprocessorDefinitions>%(PreprocessorDefinitions)</PreprocessorDefinitions>
    </ClCompile>
  </ItemDefinitionGroup>

In this case, I was missing NOMINMAX. I'm not sure why, but the
functionality that was being used to get these values just left it out.

The tool wrappers can each cast their wrapped objects to an
IVCRulePropertyStorage, which does return the correct information. It also
handles property sheets and everything else, so it should simplify the
code a lot.

Fixes #12