Updated Readme with troubleshooting information
Sourcetrail may grey out the `Create Compilation Database` option if it fails
to find any C/C++ projects. It might fail to find said projects if Intellisense
is disabled. When Intellisense is disabled, DTE fails to get the Code Model
language for the project with a rather non-obvious error about the database
being read only.

The readme now contains this information